Renamed burnoutCell -> burnout_cell
