+++ src/burnout_cell.c
@@ -1,4 +1,4 @@
-#include "burnoutCell.h"
+#include "burnout_cell.h"
 
 void burnoutCells(CellularAutomaton* automaton) {
